refactor: auth util for mocking
