+++ app/src/store/user.ts
@@ -1,8 +1,6 @@
-import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
-import { getEnvVariable } from '../util/env';
-import Auth0, { type SaveCredentialsParams } from 'react-native-auth0';
-import { State, Dispatch, type AsyncStoreSlice } from './app';
-import { settlePromise } from '../util/settlePromise';
+import { createSlice } from '@reduxjs/toolkit';
+import { type AsyncStoreSlice } from './app';
+import * as Auth from '../util/auth';
 
 export type UserState = AsyncStoreSlice<
   {
@@ -26,23 +24,23 @@ export const userSlice = createSlice({
   reducers: {},
   extraReducers: (builder) => {
     builder
-      .addCase(checkLoginStatus.pending, () => {
+      .addCase(Auth.checkLoginStatus.pending, () => {
         return { status: 'pending' };
       })
-      .addCase(checkLoginStatus.fulfilled, (_, action) => {
+      .addCase(Auth.checkLoginStatus.fulfilled, (_, action) => {
         return { status: 'fulfilled', ...action.payload };
       })
-      .addCase(checkLoginStatus.rejected, () => {
+      .addCase(Auth.checkLoginStatus.rejected, () => {
         return { status: 'rejected', reason: 'CHECK_LOGIN_STATUS' };
       });
     builder
-      .addCase(login.pending, () => {
+      .addCase(Auth.login.pending, () => {
         return { status: 'pending' };
       })
-      .addCase(login.fulfilled, (_, action) => {
+      .addCase(Auth.login.fulfilled, (_, action) => {
         return { status: 'fulfilled', ...action.payload };
       })
-      .addCase(login.rejected, (_, action) => {
+      .addCase(Auth.login.rejected, (_, action) => {
         return {
           status: 'rejected',
           reason: action.payload?.reason,
@@ -50,13 +48,13 @@ export const userSlice = createSlice({
         };
       });
     builder
-      .addCase(logout.pending, () => {
+      .addCase(Auth.logout.pending, () => {
         return { status: 'pending' };
       })
-      .addCase(logout.fulfilled, (_, action) => {
+      .addCase(Auth.logout.fulfilled, (_, action) => {
         return { status: 'fulfilled', ...action.payload };
       })
-      .addCase(logout.rejected, (_, action) => {
+      .addCase(Auth.logout.rejected, (_, action) => {
         return {
           status: 'rejected',
           reason: action.payload?.reason,
@@ -64,92 +62,4 @@ export const userSlice = createSlice({
         };
       });
   },
-});
-
-const auth0 = new Auth0({
-  domain: getEnvVariable('DOMAIN_NAME'),
-  clientId: getEnvVariable('CLIENT_ID'),
-});
-
-export const checkLoginStatus = createAsyncThunk<
-  {
-    accessToken: string;
-    email: string;
-  },
-  void,
-  {
-    state: State;
-    dispatch: Dispatch;
-    rejectValue: undefined;
-  }
->('user/checkLoginStatus', async () => {
-  const getCredentialsResult = await settlePromise(
-    auth0.credentialsManager.getCredentials(),
-  );
-
-  if (getCredentialsResult.status === 'fulfilled') {
-    const credentials = getCredentialsResult.value;
-    const user = await auth0.auth.userInfo({
-      token: credentials.accessToken,
-    });
-    return {
-      accessToken: credentials.accessToken,
-      email: user.email,
-    };
-  } else {
-    return { accessToken: '', email: '' };
-  }
-});
-
-export const login = createAsyncThunk<
-  { accessToken: string; email: string },
-  void,
-  {
-    state: State;
-    dispatch: Dispatch;
-    rejectValue: { reason: 'ID_TOKEN_UNDEFINED' | 'LOGIN_ERROR'; error?: any };
-  }
->('user/login', async (_, { rejectWithValue }) => {
-  try {
-    const credentials = await auth0.webAuth.authorize({
-      scope: getEnvVariable('AUTH0_SCOPE'),
-      audience: getEnvVariable('AUTH0_AUDIENCE'),
-    });
-
-    if (typeof credentials.idToken === 'undefined') {
-      return rejectWithValue({ reason: 'ID_TOKEN_UNDEFINED' });
-    }
-
-    auth0.credentialsManager.saveCredentials(
-      credentials as SaveCredentialsParams,
-    );
-
-    const user = await auth0.auth.userInfo({
-      token: credentials.accessToken,
-    });
-    return {
-      accessToken: credentials.accessToken,
-      email: user.email,
-    };
-  } catch (error) {
-    return rejectWithValue({ reason: 'LOGIN_ERROR', error });
-  }
-});
-
-export const logout = createAsyncThunk<
-  { accessToken: string; email: string },
-  void,
-  {
-    state: State;
-    dispatch: Dispatch;
-    rejectValue: { reason: 'LOGOUT_ERROR'; error: any };
-  }
->('user/logout', async (_, { rejectWithValue }) => {
-  try {
-    await auth0.webAuth.clearSession();
-    await auth0.credentialsManager.clearCredentials();
-    return { accessToken: '', email: '' };
-  } catch (error) {
-    return rejectWithValue({ reason: 'LOGOUT_ERROR', error });
-  }
 });
